refactor(ai_chat): ストリーミング実体を ai_chat_service へ分離 (#782 R5)
commands/ai_chat.rs (1,089 行) から、透過リトライ状態機械 (#646)・SSE
パース・Anthropic/OpenAI 互換 tool_use ビルダー・ストリーム台帳
(ACTIVE_STREAMS、poison 復旧込み)・secret redaction・専用 HTTP クライアント
(#508 の HTTP/1.1 固定設定) を src-tauri/src/ai_chat_service.rs へ移す。

- コマンド層は start_stream / cancel_stream への薄い委譲のみ (~30 行)
- 入力検証 (model 空・256KB 上限) と Vault 接続解決も service 側の
  start_stream に含め、query_runtime / server_info と同格の service とする
- 既存テスト (message 変換・SSE・redaction 等) は service 内へそのまま移動。
  bindings.ts は byte 一致 (型名・doc とも不変)
